Problem: script/builtins file lumps everything together
Besides having one file for disparate instructions definitions,
this will expose instructions meant for modules that might be
not enabled during runtime (for example, if one disables mod_json,
there shouldn't be JSON/EMPTY anymore)

Solution: split builtins into per-module builtins and handle
standard builtins-related functionality in designated macros.
