+++ pumpkindb_engine/src/script/mod_binaries.builtins
@@ -1,0 +1,1 @@
+STARTSWITH? : 2DUP LENGTH SWAP LENGTH SWAP UINT/LT? [2DROP 0] [DUP LENGTH 0 SWAP 2SWAP SWAP 2SWAP SLICE EQUAL?] IFELSE.
