+++ pumpkindb_engine/src/script/mod_binaries.rs
@@ -25,8 +25,11 @@ pub struct Handler<'a> {
     phantom: PhantomData<&'a ()>,
 }
 
+builtins!("mod_binaries.builtins");
+
 impl<'a> Dispatcher<'a> for Handler<'a> {
     fn handle(&mut self, env: &mut Env<'a>, instruction: &'a [u8], pid: EnvId) -> PassResult<'a> {
+        try_instruction!(env, self.handle_builtins(env, instruction, pid));
         try_instruction!(env, self.handle_ltp(env, instruction, pid));
         try_instruction!(env, self.handle_gtp(env, instruction, pid));
         try_instruction!(env, self.handle_equal(env, instruction, pid));
@@ -42,6 +45,8 @@ impl<'a> Handler<'a> {
     pub fn new() -> Self {
         Handler { phantom: PhantomData }
     }
+
+    handle_builtins!();
 
     #[inline]
     fn handle_equal(&mut self,
