+++ pumpkindb_engine/src/script/mod_core.builtins
@@ -1,0 +1,8 @@
+( Constants )
+$SYSTEM/VERSION : "0.2-wip".
+TRUE : 0x01.
+FALSE : 0x00.
+( Basic built-ins )
+IF : [] IFELSE.
+NONE? : LENGTH 0 EQUAL?.
+SOME? : NONE? NOT.
