+++ pumpkindb_engine/src/script/mod_json.builtins
@@ -1,0 +1,1 @@
+JSON/EMPTY : "{}" .
