+++ pumpkindb_engine/src/script/mod_stack.builtins
@@ -1,0 +1,8 @@
+2DUP : OVER OVER.
+2DROP : DROP DROP.
+STACK : DEPTH WRAP.
+NIP : SWAP DROP.
+TUCK : SWAP OVER.
+2NIP : 2SWAP 2DROP.
+2TUCK : 2SWAP 2OVER.
+-ROT : ROT ROT.
